[Task2] Finished project.

+++ api-core/src/main/java/com/illiapinchuk/testtask/service/impl/AuctionServiceImpl.java
@@ -2,6 +2,7 @@ package com.illiapinchuk.testtask.service.impl;
 
 import com.illiapinchuk.testtask.common.mapper.AuctionMapper;
 import com.illiapinchuk.testtask.common.validator.AuctionValidator;
+import com.illiapinchuk.testtask.configuration.security.UserPermissionService;
 import com.illiapinchuk.testtask.exception.AuctionAlreadyExistsException;
 import com.illiapinchuk.testtask.exception.AuctionNotFoundException;
 import com.illiapinchuk.testtask.exception.UserIsNotOwnerOfTheAuctionException;
@@ -9,6 +10,7 @@ import com.illiapinchuk.testtask.model.dto.AuctionDto;
 import com.illiapinchuk.testtask.persistence.entity.Auction;
 import com.illiapinchuk.testtask.persistence.repository.AuctionRepository;
 import com.illiapinchuk.testtask.service.AuctionService;
+import com.illiapinchuk.testtask.service.UserService;
 import jakarta.annotation.Nonnull;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -22,6 +24,8 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 @Slf4j
 public class AuctionServiceImpl implements AuctionService {
+
+  private final UserService userService;
 
   private final AuctionRepository auctionRepository;
 
@@ -48,9 +52,16 @@ public class AuctionServiceImpl implements AuctionService {
   @Override
   @Transactional
   public Auction createAuction(@Nonnull final Auction auction) {
-    if (auction.getId() != null && auctionValidator.isAuctionExistsInDbById(auction.getId()))
+    final var currentUser = userService.getUserById(UserPermissionService.getJwtUser().getId());
+
+    // Check if the auction already exists
+    if (auction.getId() != null && auctionValidator.isAuctionExistsInDbById(auction.getId())) {
       throw new AuctionAlreadyExistsException(
           String.format("Auction with id: %s already exists", auction.getId()));
+    }
+
+    // Set the creator of the auction
+    auction.setCreator(currentUser);
 
     return auctionRepository.save(auction);
   }
@@ -61,7 +72,7 @@ public class AuctionServiceImpl implements AuctionService {
     final var auctionId = auctionDto.getId();
 
     checkIfAuctionExists(auctionId);
-    checkIfUserCreatedAuction(auctionId, auctionDto.getCreatorId());
+    checkIfUserHaveAccessToAuction(auctionId, auctionDto.getCreatorId());
 
     final var auction = getAuctionById(auctionId);
 
@@ -72,17 +83,21 @@ public class AuctionServiceImpl implements AuctionService {
 
   @Override
   public void deleteAuctionById(@Nonnull final Long id) {
+    checkIfUserHaveAccessToAuction(id, UserPermissionService.getJwtUser().getId());
+
     auctionRepository.deleteById(id);
   }
 
-  private void checkIfAuctionExists(Long auctionId) {
+  private void checkIfAuctionExists(@Nonnull final Long auctionId) {
     if (!auctionValidator.isAuctionExistsInDbById(auctionId)) {
       throw new AuctionNotFoundException(String.format("Auction with id: %s not found", auctionId));
     }
   }
 
-  private void checkIfUserCreatedAuction(Long auctionId, Long creatorId) {
-    if (!auctionValidator.isUserCreatedAuction(auctionId, creatorId)) {
+  private void checkIfUserHaveAccessToAuction(
+      @Nonnull final Long auctionId, @Nonnull final Long creatorId) {
+    if (!auctionValidator.isUserCreatedAuction(auctionId, creatorId)
+        && !UserPermissionService.hasAnyRulingRole()) {
       throw new UserIsNotOwnerOfTheAuctionException(
           String.format(
               "User with id: %s did not create the auction with id: %s", creatorId, auctionId));
